Prevent app crash on adding invalid stock symbol
Adding an invalid stock symbol throws a NullPointerException in
QuoteSyncJob. On catching this exception, an error event with
appropriate error code is published. The subscriber StockListFragment
subscribes to this event and updates user regarding the same.

Two methods have been tried to notify StockListFragment of error. First
was using broadcasts, and the second an eventBus. Retaining both
methods, however currently using only eventBus.