Navigate to StockDetailActivity on clicking a stock in list view
Refactored StockAdapterOnClickHandler to include method onStockClick
with parameters stock symbol and stock name. StockListFragment
implements this interface wherein it just delegates the call to
MainActivity if it finds an instance of it. MainActivity then starts
StockDetailActivity.

While StockAdapter itself could have started StockDetailActivity from
the reference to the context it holds, the deliberate usage of the
logic is to allow handling UI for tablets in future.