Update logic to populate stock quotes list
Switched from using loader to populating stock quote recycler view on
receiving DataUpdatedEvent. Stock quotes are loaded asynchronously
using RxJava Single.

The reason to switch is strongly biased towards finally getting started
to use RxJava in Android. It is easier to retain an ArrayList of
parcelable objects in saveInstanceState. The cursor was never closed
either so had to find out if that caused any memory leaks.